test_env_setup_util/requirements.txt
Jinja2==3.1.6
jsonschema==4.23.0
paramiko==3.5.1
PyYAML==6.0.2
scp==0.15.0